2003-11-03  Bharat Mediratta  <bharat@menalto.com> 1.4.1-cvs-b206
        * Work around a PHP bug on OpenBSD where if you store the results of
          ceil() in a session variable, it goes in as a float (eg "1.0000")
          instead of an int, triggering OpenBSD's serialization bug


git-svn-id: file:///home/jmullan/src/gallery-git/repo/trunk/gallery@4593 57fcd75e-5312-0410-8df3-f5eb6fbb1595

+++ view_photo.php
@@ -146,7 +146,7 @@ if ($index > $gallery->album->numPhotos(1)) {
 $rows = $gallery->album->fields["rows"];
 $cols = $gallery->album->fields["cols"];
 $perPage = $rows * $cols;
-$page = ceil($index / ($rows * $cols));
+$page = (int)(ceil($index / ($rows * $cols)));
 
 $gallery->session->albumPage[$gallery->album->fields['name']] = $page;
